Added vcpp project to the dh_branch. also minor modifications in the code to compile on windows.
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/branches/dh_branch@3271 d1858a74-f715-0410-99c9-c66fe9cce3d6

Former-commit-id: 6e5836694178815a5da6227112657eabd87626b3
Former-commit-id: f50ff39ed5a8fa533cfe1f2234a65816089ee806

+++ builtins/testBuiltins.cpp
@@ -269,7 +269,7 @@ void testTable()
 	Id tabentry( tabid.value() + 1 );
 	Table* t = reinterpret_cast< Table* >( tabid.eref().data() );
 	for ( unsigned int i = 0; i < 100; ++i ) {
-		t->input( sqrt( i ) );
+		t->input( sqrt((double) i ) );
 	}
 	unsigned int numEntries = Field< unsigned int >::get( 
 		tabid, "num_table" );
@@ -282,7 +282,7 @@ void testTable()
 	for ( unsigned int i = 0; i < 100; ++i ) {
 		ObjId temp( tabentry, DataId( i ) );
 		double ret = Field< double >::get( temp, "value" );
-		assert( fabs( ret - sqrt( i ) ) < 1e-6 );
+		assert( fabs( ret - sqrt((double) i ) ) < 1e-6 );
 	}
 	/*
 	SetGet2< string, string >::set(
